requirements.txt
faiss-cpu>=1.7.4
gensim>=4.3.2
nltk>=3.8.1
scikit-learn>=1.3.0
numpy>=1.24.0
transformers>=4.40.0
torch>=2.0.0
requests>=2.31.0
streamlit>=1.30.0
pymupdf>=1.23.0
sentence-transformers>=2.2.2